feat: Add setOrg function to useAppContext hook in HeaderV0 component
The code changes in HeaderV0 component add the setOrg function to the useAppContext hook. This allows for setting the organization name when a repository is selected. The setOrg function is called when a repository is selected, passing the org value from the selected repository. If no repository is selected, the org value is set to an empty string.